Move Value-related stuff under semantics::nze
