--- dhall/src/semantics/core/mod.rs
@@ -1,6 +1,0 @@
-pub mod context;
-pub mod value;
-pub mod var;
-pub mod visitor;
-pub(crate) use value::*;
-pub(crate) use var::*;
