+++ dhall/src/semantics/nze/mod.rs
@@ -1,2 +1,7 @@
 pub mod env;
+pub mod value;
+pub mod var;
+pub mod visitor;
 pub(crate) use env::*;
+pub(crate) use value::*;
+pub(crate) use var::*;
